+++ src/wai/annotations/domain/audio/classification/_ClassificationInstance.py
@@ -1,0 +1,13 @@
+from typing import Type
+
+from .._AudioInstance import AudioInstance
+from ...classification import Classification
+
+
+class ClassificationInstance(AudioInstance[Classification]):
+    """
+    An instance in the audio classification domain.
+    """
+    @classmethod
+    def annotations_type(cls) -> Type[Classification]:
+        return Classification