+++ src/wai/annotations/domain/audio/classification/__init__.py
@@ -1,0 +1,2 @@
+from ._ClassificationDomainSpecifier import ClassificationDomainSpecifier
+from ._ClassificationInstance import ClassificationInstance