+++ src/wai/annotations/format/audio/component/__init__.py
@@ -1,1 +1,2 @@
+from ._AudioAC import AudioReaderAC, AudioWriterAC
 from ._AudioSP import AudioReaderSP, AudioWriterSP