+++ src/wai/annotations/format/audio/specifier/__init__.py
@@ -1,2 +1,4 @@
+from ._AudioACInputFormatSpecifier import AudioACInputFormatSpecifier
+from ._AudioACOutputFormatSpecifier import AudioACOutputFormatSpecifier
 from ._AudioSPInputFormatSpecifier import AudioSPInputFormatSpecifier
 from ._AudioSPOutputFormatSpecifier import AudioSPOutputFormatSpecifier