+++ CHANGES.rst
@@ -4,8 +4,9 @@ Changelog
 0.1.8 (???-??-??)
 -----------------
 
-- Added dataset reader for audio files: `from-audio-file-sp`
-- Added dataset writer for audio files: `to-audio-file-sp`
+- Added new audio domain for classification using suffix `-ac`
+- Added dataset reader for audio files: `from-audio-files-sp`, `from-audio-files-ac`
+- Added dataset writer for audio files: `to-audio-files-sp`, `to-audio-files-ac`
 
 
 0.1.7 (2022-06-13)